<refactor>(gaan_utils.py)<rename>
